Add 'RZ' to the geometryType_t coordinates enum.
This proves very useful to permit the templating of kernels to maintain execution efficiency when run in an [Nr 1 Nz] simulation,
regardless of coordinates (cylindrical or square). Very large increases in efficiency of several source kernels were acheived, and
the speed of the HLL flux kernel was improved roughly 25%.

The scalar potential kernel's efficiency was improved further by decomposing it into the gradient-calculation and the force-application
part; This removes the force-apply, which is trivially parallel given the gradient, from the gradient-calculate which requires a
limited degree of parallel synchronization.